Vectorise every evaluation route (audit PR 9)
The tuple-vectorisation principle was honoured in shape and not in cost:
array evaluation points returned correctly shaped answers while per-point
Python loops ran underneath. Three causes are repaired here.

Priors rebuilt a frozen SciPy distribution on every density call, so 79% of
the default route's runtime was inside rv_frozen.__init__. The symbolic
backend substituted point by point, which was 97.6% of the scipy route;
it now compiles with lambdify, with a probe at setup because expint (used
by the Pareto MGF) compiles but does not run, and with an exact-path
fallback for any element that overflows, underflows or returns NaN. The
quadrature ran once per point and now batches onto a common interval.

Per-point cost at a batch of eight: scipy 2054.5 -> 2.0 ms, symbolic
integer 1.1 -> 0.02 ms, auto fractional 255.9 -> 9.7 ms, jax 657.7 -> 66.6
ms, bell 8.5 -> 1.7 ms, auto integer 24.0 -> 7.8 ms. Accuracy improved from
a worst case of 1.17e-11 to 7.85e-15. The full suite went from 623 s to 67 s.

The array-valued-order path is unchanged and still flat, because different
orders may resolve to different backends and cannot share a call.

+++ jumufraktiv/numeric_fractionalDeriv_grid.py
@@ -251,7 +251,7 @@ def fractionalDeriv_grid(
     # ---- Fixed grid, with the right endpoint found by decay probing -----
     u_min, u_max = _integration_range(gamma_val, tol)
     while True:
-        n_nodes = max(int(math.ceil((u_max - u_min) * _NODES_PER_UNIT)) + 1, 33)
+        n_nodes = max(math.ceil((u_max - u_min) * _NODES_PER_UNIT) + 1, 33)
         nodes, step = np.linspace(u_min, u_max, n_nodes, retstep=True)
         decayed, log_edge, log_peak = _tail_has_decayed(
             nodes, gamma_val, t_arr, u_arr, integer_derivative, tol
